requirements.txt
-f http://bdc-pypi.s3-website-ap-northeast-1.amazonaws.com/bdc
--trusted-host=bdc-pypi.s3-website-ap-northeast-1.amazonaws.com
bdc==0.1.78

ipython
datasets>=2.0.0
einops
jsonlines
numexpr
openai>=0.6.4
omegaconf>=2.2
peft>=0.2.0
pybind11>=2.6.2
pycountry
pytablewriter
rouge-score>=0.0.4
sacrebleu==1.5.0
scikit-learn>=0.24.1
sqlitedict
torch<=2.0.1
tqdm-multiprocess
transformers>=4.1
zstandard
accelerate>=0.17.1
python-dotenv